+++ CHANGELOG.md
@@ -47,6 +47,7 @@ high state of flux, you're at risk of it changing without notice.
     - rename `ReadonlyNonEmptyArray` module to `NonEmptyArray` (@gcanti)
     - rename `ReadonlyRecord` module to `Record` (@gcanti)
     - rename `ReadonlySet` module to `Set` (@gcanti)
+    - rename `ReadonlyTuple` module to `Tuple` (@gcanti)
     - `Applicative`
       - remove `getApplicativeComposition` and related types (@gcanti)
     - `Compactable`